ov072: rename SnowmanBody to daBgSnmBdy_c and promote the 29-function TU
The cartridge names this class. ov072 0x0212278c holds the bytes
"12daBgSnmBdy_c" NUL-terminated -- the length-prefixed Itanium _ZTS payload --
_ZTI12daBgSnmBdy_c at 0x02122780 points its +4 word at it, and the vtable's -4
header word at 0x021227bc points back at that _ZTI. "SnowmanBody" was a coined
readable spelling; ov072/symbols.txt already carried _ZTI12daBgSnmBdy_c,
_ZTS12daBgSnmBdy_c and daBgSnmBdy_c_classInit under the cartridge's own name,
and only the vtable address carried both spellings.

The rename and the promotion are one change because dsd resolves .data records
by NAME: a half-landed rename links the vtable word the factory stores as
0x00000000. The coined alias row _ZTV11SnowmanBody at 0x021227c0 is dropped so
exactly one spelling survives at that address; the address itself is unchanged.

  include/SnowmanBody.h -> include/daBgSnmBdy_c.h, struct SnowmanBody ->
  struct daBgSnmBdy_c, 28 src/_ZN11SnowmanBody*.cpp shards plus
  src/d_a_bg_snm_bdy.c folded into src/actors/daBgSnmBdy_c.cpp.

THE FOLD IS 29, NOT 28. build/tu_map.json stops the run at 0x0211fedc, so the
queue's shard_count is a floor. The registry maps BIG_SNOWMAN_BODY to the
0x0212279c descriptor whose first word relocates to daBgSnmBdy_c_classInit at
0x0211fedc; that factory allocates 0x3a8 through fBase_c::operator new,
installs this vtable, constructs the five typed subobjects and ends exactly at
daBgSnmHed_c's D1 at 0x0211ff34. Absorbing it makes the run 29 functions,
0x0211f000..0x0211ff34, with no gap.

FORM, copied from the promoted in-overlay sibling ov072/daBgSnwmn_c:

  - the destructor is inline and declared LAST in the header, so mwccarm emits
    D1 then D0 at the bottom of the run and no D2 body -- the cartridge's own
    order;
  - a class-local operator new routes `new daBgSnmBdy_c()` to the retail
    fBase_c allocator, so the COMPILER emits and owns the vptr store and no
    _ZTV address is spelled by hand anywhere in the source;
  - the actor profile is emitted from a typed 0x1c struct;
  - the same STB_LOPROC -> STB_GLOBAL destructor binding rewrites;
  - the same six inherited vague-linkage RTTI records are externalized to
    their arm9 canonical addresses.

InitResources is the key function -- the first out-of-line virtual the class
declares once the destructor is inline -- so this TU owns _ZTV12daBgSnmBdy_c,
_ZTI12daBgSnmBdy_c and _ZTS12daBgSnmBdy_c. The route is intact-object: the TU
claims .text 0x0211f000..0x0211ff34 and .data 0x02122780..0x0212283c.

PROOF at this tree, base a278141a9:

  - tubuild.py verify ov072/daBgSnmBdy_c: 29/29 MATCH, objisolate clean,
    relocation destinations clean, all 29 sections in ROM-ascending order.
  - tubuild.py linkcheck ov072/daBgSnmBdy_c: SCRATCH-DATA-VERIFIED. Both
    licensed ranges reproduce with 0 differing bytes, 33 LICENSED symbols and
    no unlicensed section, dsd check symbols has 0 new errors against the
    refreshed baseline control's 9 pre-existing ones, and the scratch ROM is
    byte-identical to stock.
  - rombuild.py -j16: 106/106 modules exact, 11,191 source-built functions
    reproducing and 0 mismatching, 21 source-owned data claims reproducing,
    and build/sm64ds.nds sha256
    d1506e90efae5e2d2cf119926a4ac2a291bd5ca78349d09d5024e1a918c478e8, the
    stock ROM. ROM data verified from source goes 687 -> 689 symbols and
    44,288 -> 44,328 bytes against the same build at the base; differing stays
    at 5.

"Promoted" is a packaging state. Twenty-two of the twenty-nine members still
reach the object through raw byte offsets rather than the named fields the
header declares, and the TU still spells other modules' mangled symbols
directly. That is remaining reconstruction work, not a byte defect.
