+++ attribution.json
@@ -1001,6 +1001,35 @@
     "src/game/actors/d_a_bg_snwmn.cpp#_ZN11daBgSnwmn_cD1Ev": "github-actions[bot]",
     "src/game/actors/d_a_bg_snwmn.cpp#daBgSnwmn_c_classInit": "github-actions[bot]",
     "src/game/actors/d_a_bg_snwmn.cpp#func_ov072_021208d8": "github-actions[bot]",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_cD1Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_cD0Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c18IsPlayerNearCenterEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c15UpdateRollAngleEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c11AdvancePathEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c10HurtPlayerEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c21UpdateGroundCollisionEP10dBgCh_Actr": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c11UpdateModelEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6State5Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c10InitState5Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6State4Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c10InitState4Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6State3Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c10InitState3Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6State2Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c10InitState2Ev": "tangosdev",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6State1Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c10InitState1Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6State0Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c10InitState0Ev": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c17CallStateBehaviorEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c13CallStateInitEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c8SetStateEi": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c16CleanupResourcesEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c16OnPendingDestroyEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6RenderEv": "andrewboudreau",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c8BehaviorEv": "tangosdev",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c13InitResourcesEv": "tangosdev",
+    "src/actors/daBgSnmBdy_c.cpp#daBgSnmBdy_c_classInit": "tangosdev",
     "src/game/actors/d_a_cam_tag.cpp#_ZN10daCamTag_c13InitResourcesEv": "github-actions[bot]",
     "src/game/actors/d_a_cam_tag.cpp#_ZN10daCamTag_c16CleanupResourcesEv": "github-actions[bot]",
     "src/game/actors/d_a_cam_tag.cpp#_ZN10daCamTag_c16OnPendingDestroyEv": "andrewboudreau",
