+++ config/arm9/overlays/ov072/delinks.txt
@@ -4,121 +4,10 @@
     .ctor       start:0x02122708 end:0x02122718 kind:rodata align:4
     .data       start:0x02122720 end:0x02122b20 kind:data align:32
     .bss        start:0x02122b20 end:0x02122de0 kind:bss align:32
-src/_ZN11SnowmanBodyD1Ev.cpp:
+src/actors/daBgSnmBdy_c.cpp:
     complete
-    .text start:0x0211f000 end:0x0211f048
-
-src/_ZN11SnowmanBodyD0Ev.cpp:
-    complete
-    .text start:0x0211f048 end:0x0211f0a4
-
-src/_ZN11SnowmanBody18IsPlayerNearCenterEv.cpp:
-    complete
-    .text start:0x0211f0a4 end:0x0211f158
-
-src/_ZN11SnowmanBody15UpdateRollAngleEv.cpp:
-    complete
-    .text start:0x0211f158 end:0x0211f1dc
-
-src/_ZN11SnowmanBody11AdvancePathEv.cpp:
-    complete
-    .text start:0x0211f1dc end:0x0211f280
-
-src/_ZN11SnowmanBody10HurtPlayerEv.cpp:
-    complete
-    .text start:0x0211f280 end:0x0211f330
-
-src/_ZN11SnowmanBody21UpdateGroundCollisionEP10dBgCh_Actr.cpp:
-    complete
-    .text start:0x0211f330 end:0x0211f3e4
-
-src/_ZN11SnowmanBody11UpdateModelEv.cpp:
-    complete
-    .text start:0x0211f3e4 end:0x0211f48c
-
-src/_ZN11SnowmanBody6State5Ev.cpp:
-    complete
-    .text start:0x0211f48c end:0x0211f578
-
-src/_ZN11SnowmanBody10InitState5Ev.cpp:
-    complete
-    .text start:0x0211f578 end:0x0211f598
-
-src/_ZN11SnowmanBody6State4Ev.cpp:
-    complete
-    .text start:0x0211f598 end:0x0211f63c
-
-src/_ZN11SnowmanBody10InitState4Ev.cpp:
-    complete
-    .text start:0x0211f63c end:0x0211f65c
-
-src/_ZN11SnowmanBody6State3Ev.cpp:
-    complete
-    .text start:0x0211f65c end:0x0211f804
-
-src/_ZN11SnowmanBody10InitState3Ev.cpp:
-    complete
-    .text start:0x0211f804 end:0x0211f81c
-
-src/_ZN11SnowmanBody6State2Ev.cpp:
-    complete
-    .text start:0x0211f81c end:0x0211f9c4
-
-src/_ZN11SnowmanBody10InitState2Ev.cpp:
-    complete
-    .text start:0x0211f9c4 end:0x0211fa08
-
-src/_ZN11SnowmanBody6State1Ev.cpp:
-    complete
-    .text start:0x0211fa08 end:0x0211faf0
-
-src/_ZN11SnowmanBody10InitState1Ev.cpp:
-    complete
-    .text start:0x0211faf0 end:0x0211fb14
-
-src/_ZN11SnowmanBody6State0Ev.cpp:
-    complete
-    .text start:0x0211fb14 end:0x0211fb7c
-
-src/_ZN11SnowmanBody10InitState0Ev.cpp:
-    complete
-    .text start:0x0211fb7c end:0x0211fc3c
-
-src/_ZN11SnowmanBody17CallStateBehaviorEv.cpp:
-    complete
-    .text start:0x0211fc3c end:0x0211fc78
-
-src/_ZN11SnowmanBody13CallStateInitEv.cpp:
-    complete
-    .text start:0x0211fc78 end:0x0211fcb0
-
-src/_ZN11SnowmanBody8SetStateEi.cpp:
-    complete
-    .text start:0x0211fcb0 end:0x0211fccc
-
-src/_ZN11SnowmanBody16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x0211fccc end:0x0211fcf0
-
-src/_ZN11SnowmanBody16OnPendingDestroyEv.cpp:
-    complete
-    .text start:0x0211fcf0 end:0x0211fcf4
-
-src/_ZN11SnowmanBody6RenderEv.cpp:
-    complete
-    .text start:0x0211fcf4 end:0x0211fd24
-
-src/_ZN11SnowmanBody8BehaviorEv.cpp:
-    complete
-    .text start:0x0211fd24 end:0x0211fd54
-
-src/_ZN11SnowmanBody13InitResourcesEv.cpp:
-    complete
-    .text start:0x0211fd54 end:0x0211fedc
-
-src/d_a_bg_snm_bdy.c:
-    complete
-    .text start:0x0211fedc end:0x0211ff34
+    .text start:0x0211f000 end:0x0211ff34
+    .data start:0x02122780 end:0x0212283c
 
 src/_ZN11SnowmanHeadD1Ev.cpp:
     complete
