+++ config/arm9/overlays/ov072/symbols.txt
@@ -1,31 +1,31 @@
-_ZN11SnowmanBodyD1Ev kind:function(arm,size=0x48) addr:0x0211f000
-_ZN11SnowmanBodyD0Ev kind:function(arm,size=0x5c) addr:0x0211f048
-_ZN11SnowmanBody18IsPlayerNearCenterEv kind:function(arm,size=0xb4) addr:0x0211f0a4
-_ZN11SnowmanBody15UpdateRollAngleEv kind:function(arm,size=0x84) addr:0x0211f158
-_ZN11SnowmanBody11AdvancePathEv kind:function(arm,size=0xa4) addr:0x0211f1dc
-_ZN11SnowmanBody10HurtPlayerEv kind:function(arm,size=0xb0) addr:0x0211f280
-_ZN11SnowmanBody21UpdateGroundCollisionEP10dBgCh_Actr kind:function(arm,size=0xb4) addr:0x0211f330
-_ZN11SnowmanBody11UpdateModelEv kind:function(arm,size=0xa8) addr:0x0211f3e4
-_ZN11SnowmanBody6State5Ev kind:function(arm,size=0xec) addr:0x0211f48c
-_ZN11SnowmanBody10InitState5Ev kind:function(arm,size=0x20) addr:0x0211f578
-_ZN11SnowmanBody6State4Ev kind:function(arm,size=0xa4) addr:0x0211f598
-_ZN11SnowmanBody10InitState4Ev kind:function(arm,size=0x20) addr:0x0211f63c
-_ZN11SnowmanBody6State3Ev kind:function(arm,size=0x1a8) addr:0x0211f65c
-_ZN11SnowmanBody10InitState3Ev kind:function(arm,size=0x18) addr:0x0211f804
-_ZN11SnowmanBody6State2Ev kind:function(arm,size=0x1a8) addr:0x0211f81c
-_ZN11SnowmanBody10InitState2Ev kind:function(arm,size=0x44) addr:0x0211f9c4
-_ZN11SnowmanBody6State1Ev kind:function(arm,size=0xe8) addr:0x0211fa08
-_ZN11SnowmanBody10InitState1Ev kind:function(arm,size=0x24) addr:0x0211faf0
-_ZN11SnowmanBody6State0Ev kind:function(arm,size=0x68) addr:0x0211fb14
-_ZN11SnowmanBody10InitState0Ev kind:function(arm,size=0xc0) addr:0x0211fb7c
-_ZN11SnowmanBody17CallStateBehaviorEv kind:function(arm,size=0x3c) addr:0x0211fc3c
-_ZN11SnowmanBody13CallStateInitEv kind:function(arm,size=0x38) addr:0x0211fc78
-_ZN11SnowmanBody8SetStateEi kind:function(arm,size=0x1c) addr:0x0211fcb0
-_ZN11SnowmanBody16CleanupResourcesEv kind:function(arm,size=0x24) addr:0x0211fccc
-_ZN11SnowmanBody16OnPendingDestroyEv kind:function(arm,size=0x4) addr:0x0211fcf0
-_ZN11SnowmanBody6RenderEv kind:function(arm,size=0x30) addr:0x0211fcf4
-_ZN11SnowmanBody8BehaviorEv kind:function(arm,size=0x30) addr:0x0211fd24
-_ZN11SnowmanBody13InitResourcesEv kind:function(arm,size=0x188) addr:0x0211fd54
+_ZN12daBgSnmBdy_cD1Ev kind:function(arm,size=0x48) addr:0x0211f000
+_ZN12daBgSnmBdy_cD0Ev kind:function(arm,size=0x5c) addr:0x0211f048
+_ZN12daBgSnmBdy_c18IsPlayerNearCenterEv kind:function(arm,size=0xb4) addr:0x0211f0a4
+_ZN12daBgSnmBdy_c15UpdateRollAngleEv kind:function(arm,size=0x84) addr:0x0211f158
+_ZN12daBgSnmBdy_c11AdvancePathEv kind:function(arm,size=0xa4) addr:0x0211f1dc
+_ZN12daBgSnmBdy_c10HurtPlayerEv kind:function(arm,size=0xb0) addr:0x0211f280
+_ZN12daBgSnmBdy_c21UpdateGroundCollisionEP10dBgCh_Actr kind:function(arm,size=0xb4) addr:0x0211f330
+_ZN12daBgSnmBdy_c11UpdateModelEv kind:function(arm,size=0xa8) addr:0x0211f3e4
+_ZN12daBgSnmBdy_c6State5Ev kind:function(arm,size=0xec) addr:0x0211f48c
+_ZN12daBgSnmBdy_c10InitState5Ev kind:function(arm,size=0x20) addr:0x0211f578
+_ZN12daBgSnmBdy_c6State4Ev kind:function(arm,size=0xa4) addr:0x0211f598
+_ZN12daBgSnmBdy_c10InitState4Ev kind:function(arm,size=0x20) addr:0x0211f63c
+_ZN12daBgSnmBdy_c6State3Ev kind:function(arm,size=0x1a8) addr:0x0211f65c
+_ZN12daBgSnmBdy_c10InitState3Ev kind:function(arm,size=0x18) addr:0x0211f804
+_ZN12daBgSnmBdy_c6State2Ev kind:function(arm,size=0x1a8) addr:0x0211f81c
+_ZN12daBgSnmBdy_c10InitState2Ev kind:function(arm,size=0x44) addr:0x0211f9c4
+_ZN12daBgSnmBdy_c6State1Ev kind:function(arm,size=0xe8) addr:0x0211fa08
+_ZN12daBgSnmBdy_c10InitState1Ev kind:function(arm,size=0x24) addr:0x0211faf0
+_ZN12daBgSnmBdy_c6State0Ev kind:function(arm,size=0x68) addr:0x0211fb14
+_ZN12daBgSnmBdy_c10InitState0Ev kind:function(arm,size=0xc0) addr:0x0211fb7c
+_ZN12daBgSnmBdy_c17CallStateBehaviorEv kind:function(arm,size=0x3c) addr:0x0211fc3c
+_ZN12daBgSnmBdy_c13CallStateInitEv kind:function(arm,size=0x38) addr:0x0211fc78
+_ZN12daBgSnmBdy_c8SetStateEi kind:function(arm,size=0x1c) addr:0x0211fcb0
+_ZN12daBgSnmBdy_c16CleanupResourcesEv kind:function(arm,size=0x24) addr:0x0211fccc
+_ZN12daBgSnmBdy_c16OnPendingDestroyEv kind:function(arm,size=0x4) addr:0x0211fcf0
+_ZN12daBgSnmBdy_c6RenderEv kind:function(arm,size=0x30) addr:0x0211fcf4
+_ZN12daBgSnmBdy_c8BehaviorEv kind:function(arm,size=0x30) addr:0x0211fd24
+_ZN12daBgSnmBdy_c13InitResourcesEv kind:function(arm,size=0x188) addr:0x0211fd54
 daBgSnmBdy_c_classInit kind:function(arm,size=0x58) addr:0x0211fedc
 _ZN11SnowmanHeadD1Ev kind:function(arm,size=0x48) addr:0x0211ff34
 _ZN11SnowmanHeadD0Ev kind:function(arm,size=0x5c) addr:0x0211ff7c
@@ -113,7 +113,6 @@ data_ov072_02122778 kind:data(any) addr:0x02122778
 _ZTI12daBgSnmBdy_c kind:data(any) addr:0x02122780
 _ZTS12daBgSnmBdy_c kind:data(any) addr:0x0212278c
 g_profile_BIG_SNOWMAN_BODY kind:data(any) addr:0x0212279c ambiguous
-_ZTV11SnowmanBody kind:data(any) addr:0x021227c0
 _ZTV12daBgSnmBdy_c kind:data(any) addr:0x021227c0
 data_ov072_0212283c kind:data(any) addr:0x0212283c
 data_ov072_02122844 kind:data(any) addr:0x02122844
