+++ config/converted-baseline.json
@@ -452,16 +452,16 @@
     "src/_ZN11ShadowModelC1Ev.cpp",
     "src/_ZN11ShadowModelD0Ev.cpp",
     "src/_ZN11ShadowModelD1Ev.cpp",
-    "src/_ZN11SnowmanBody13CallStateInitEv.cpp",
-    "src/_ZN11SnowmanBody16CleanupResourcesEv.cpp",
-    "src/_ZN11SnowmanBody16OnPendingDestroyEv.cpp",
-    "src/_ZN11SnowmanBody17CallStateBehaviorEv.cpp",
-    "src/_ZN11SnowmanBody6RenderEv.cpp",
-    "src/_ZN11SnowmanBody6State0Ev.cpp",
-    "src/_ZN11SnowmanBody8BehaviorEv.cpp",
-    "src/_ZN11SnowmanBody8SetStateEi.cpp",
-    "src/_ZN11SnowmanBodyD0Ev.cpp",
-    "src/_ZN11SnowmanBodyD1Ev.cpp",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c13CallStateInitEv",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c16CleanupResourcesEv",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c16OnPendingDestroyEv",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c17CallStateBehaviorEv",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6RenderEv",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c6State0Ev",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c8BehaviorEv",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_c8SetStateEi",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_cD0Ev",
+    "src/actors/daBgSnmBdy_c.cpp#_ZN12daBgSnmBdy_cD1Ev",
     "src/_ZN11SnowmanHead10InitState1Ev.cpp",
     "src/_ZN11SnowmanHead10InitState2Ev.cpp",
     "src/_ZN11SnowmanHead13CallStateInitEv.cpp",
