+++ include/decl_common.h
@@ -455,7 +455,6 @@ extern int _ZTV18daObjMc_Metalnet_c[];
 extern int _ZTV11CrazedCrate[];
 extern int _ZTV11PyramidStep[];
 extern int _ZTV11RollingRock[];
-extern int _ZTV11SnowmanBody[];
 extern int _ZTV11SnowmanHead[];
 extern int _ZTV11daChScene_c[];
 extern int _ZTV11WingFeather[];
