+++ notes/data/c-cpp-classification.tsv
@@ -631,9 +631,9 @@ src/_ZN11RollingRock8BehaviorEv.c	P1-vtable+mangled	ov021	daGrock_c	RollingRock/
 src/_ZN11ShadowModel9InitModelEP9Matrix4x35Fix12IiES3_S3_j.c	P3-mangled-method-only	arm9		ShadowModel	_ZN11ShadowModel9InitModelEP9Matrix4x35Fix12IiES3_S3_j	1		True		by-value-class-param
 src/_ZN11ShadowModelC1Ev.c	P3-mangled-method-only	arm9		ShadowModel	_ZN11ShadowModelC1Ev	1		True	_ZTV11ShadowModel	-
 src/_ZN11ShadowModelD0Ev.cpp	DTOR-ONLY	arm9	dExtShadowModel_c	ShadowModel/dExtShadowModel_c	_ZN11ShadowModelD0Ev	1		True	_ZTV11ShadowModel	-
-src/_ZN11SnowmanBody16CleanupResourcesEv.c	P1-vtable+mangled	ov072	daBgSnmBdy_c	SnowmanBody/daBgSnmBdy_c	vtable daBgSnmBdy_c slot 3 @0x211fccc | _ZN11SnowmanBody16CleanupResourcesEv	1		True		-
-src/_ZN11SnowmanBody16OnPendingDestroyEv.c	P1-vtable+mangled	ov072	daBgSnmBdy_c	SnowmanBody/daBgSnmBdy_c	vtable daBgSnmBdy_c slot 12 @0x211fcf0 | _ZN11SnowmanBody16OnPendingDestroyEv	1		True		-
-src/_ZN11SnowmanBodyD0Ev.c	DTOR-ONLY	ov072	daBgSnmBdy_c	SnowmanBody/daBgSnmBdy_c	_ZN11SnowmanBodyD0Ev	1		True		-
+src/_ZN12daBgSnmBdy_c16CleanupResourcesEv.c	P1-vtable+mangled	ov072	daBgSnmBdy_c	daBgSnmBdy_c/daBgSnmBdy_c	vtable daBgSnmBdy_c slot 3 @0x211fccc | _ZN12daBgSnmBdy_c16CleanupResourcesEv	1		True		-
+src/_ZN12daBgSnmBdy_c16OnPendingDestroyEv.c	P1-vtable+mangled	ov072	daBgSnmBdy_c	daBgSnmBdy_c/daBgSnmBdy_c	vtable daBgSnmBdy_c slot 12 @0x211fcf0 | _ZN12daBgSnmBdy_c16OnPendingDestroyEv	1		True		-
+src/_ZN12daBgSnmBdy_cD0Ev.c	DTOR-ONLY	ov072	daBgSnmBdy_c	daBgSnmBdy_c/daBgSnmBdy_c	_ZN12daBgSnmBdy_cD0Ev	1		True		-
 src/_ZN11SnowmanHead16CleanupResourcesEv.c	P1-vtable+mangled	ov072	daBgSnmHed_c	SnowmanHead/daBgSnmHed_c	vtable daBgSnmHed_c slot 3 @0x21205f0 | _ZN11SnowmanHead16CleanupResourcesEv	1		True		-
 src/_ZN11SnowmanHead16OnPendingDestroyEv.c	P1-vtable+mangled	ov072	daBgSnmHed_c	SnowmanHead/daBgSnmHed_c	vtable daBgSnmHed_c slot 12 @0x2120634 | _ZN11SnowmanHead16OnPendingDestroyEv	1		True		-
 src/_ZN11SnowmanHeadD0Ev.c	DTOR-ONLY	ov072	daBgSnmHed_c	SnowmanHead/daBgSnmHed_c	_ZN11SnowmanHeadD0Ev	1		True		-
