+++ notes/data/tu-promotion-queue.tsv
@@ -53,7 +53,7 @@ daPgDfdr_c	2	499	ov027	yes	yes	compiler-only:>=11(derived from own chain: 5 clas
 Rabbit	31	1694	ov085	yes	no	compiler-only:>=11(derived from own chain: 5 classes incl. self via build/rtti.json + include/ base clause; floor 2x5+1)	-	text-only
 UnknownVsEntry+UnknownVsPlayer	43	1157	ov075	partial	no	multi-class-tu:2;compiler-only:>=7(derived from own chain: 3 classes incl. self via build/rtti.json + include/ base clause; floor 2x3+1; excludes 1 unresolved: UnknownVsPlayer)	-	text-only
 daObjFloatBoard_c	3	309	ov002	yes	yes	compiler-only:>=11(derived from own chain: 5 classes incl. self via build/rtti.json; floor 2x5+1)	daBar_c	text-only
-SnowmanBody	28	665	ov072	yes	no	intact-object(owns non-text data at natural address; see sibling daBgSnwmn_c);compiler-only:>=6(derived from own chain: 4 classes incl. self via build/rtti.json + include/ base clause; floor 2x4+1-3 intact-object)	daBgSnwmn_c	intact-object
+daBgSnmBdy_c	28	665	ov072	yes	no	intact-object(owns non-text data at natural address; see sibling daBgSnwmn_c);compiler-only:>=6(derived from own chain: 4 classes incl. self via build/rtti.json + include/ base clause; floor 2x4+1-3 intact-object)	daBgSnwmn_c	intact-object
 MovingBar	23	350	ov015	yes	no	compiler-only:>=11(derived from own chain: 5 classes incl. self via build/rtti.json + include/ base clause; floor 2x5+1)	daObjBkBillboard_c	text-only
 daObjDorifu_c	6	254	ov002	yes	no	compiler-only:>=11(derived from own chain: 5 classes incl. self via build/rtti.json; floor 2x5+1)	daBar_c	text-only
 daObjGuragura_c	6	181	ov002	yes	no	compiler-only:>=11(derived from own chain: 5 classes incl. self via build/rtti.json; floor 2x5+1)	daBar_c	text-only
