--- src/_ZN11SnowmanBody13CallStateInitEv.cpp
@@ -1,8 +1,0 @@
-//cpp
-#include "SnowmanBody.h"
-
-void SnowmanBody::CallStateInit()
-{
-    StateFunc *func = mStateFuncs;
-    (this->**func)();
-}
