--- src/_ZN11SnowmanBody16OnPendingDestroyEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN11SnowmanBody16OnPendingDestroyEv
-/* SnowmanBody::OnPendingDestroy -- vtable slot 12. The ROM body is empty: the
- * override exists only to occupy the slot. */
-#include "SnowmanBody.h"
-
-void SnowmanBody::OnPendingDestroy()
-{
-}
