--- src/_ZN11SnowmanBody17CallStateBehaviorEv.cpp
@@ -1,8 +1,0 @@
-//cpp
-#include "SnowmanBody.h"
-
-void SnowmanBody::CallStateBehavior()
-{
-    StateFunc *func = mStateFuncs + 1;
-    (this->**func)();
-}
