--- src/_ZN11SnowmanBody6RenderEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN11SnowmanBody6RenderEv
-#include "SnowmanBody.h"
-
-int SnowmanBody::Render()
-{
-    mModel.Render((Vector3 *)&mScaleX);
-    return 1;
-}
