--- src/_ZN11SnowmanBodyD0Ev.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN11SnowmanBodyD0Ev
-/* Real compiler-spelled deleting destructor. dActor_c's inline operator delete
- * supplies the actor-heap release after complete SnowmanBody destruction. */
-#include "SnowmanBody.h"
-
-SnowmanBody::~SnowmanBody()
-{
-}
