--- src/_ZN11SnowmanBodyD1Ev.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN11SnowmanBodyD1Ev
-/* Real compiler-spelled complete destructor. mwccarm synthesises reverse
- * member teardown and the dActor_c base chain from SnowmanBody's layout. */
-#include "SnowmanBody.h"
-
-SnowmanBody::~SnowmanBody()
-{
-}
